requirements.txt
streamlit==1.32.0
langchain==0.1.13
langchain-openai==0.1.1
langchain-chroma==0.1.0
langchain-community==0.0.29  # Added for document loaders
python-dotenv
pypdf